+++ src/main/java/org/folio/qm/service/impl/ValidationServiceImpl.java
@@ -18,6 +18,7 @@ import org.folio.qm.exception.ValidationException;
 import org.folio.qm.service.ValidationService;
 import org.folio.qm.util.ErrorUtils;
 import org.folio.qm.validation.FieldValidationRule;
+import org.folio.qm.validation.LeaderValidationRule;
 import org.folio.qm.validation.ValidationResult;
 import org.folio.spring.FolioExecutionContext;
 
@@ -30,6 +31,7 @@ public class ValidationServiceImpl implements ValidationService {
   public static final String X_OKAPI_TOKEN_USER_ID_IS_MISSING_MESSAGE = "X-Okapi-User-Id header is missing";
 
   private final List<FieldValidationRule> fieldValidationRules;
+  private final List<LeaderValidationRule> leaderValidationRules;
 
   @Override
   public ValidationResult validate(QuickMarc quickMarc) {
@@ -42,6 +44,13 @@ public class ValidationServiceImpl implements ValidationService {
       .map(Optional::get)
       .collect(Collectors.toList());
 
+    validationErrors.addAll(leaderValidationRules.stream()
+      .filter(rule -> rule.supportFormat(marcFormat))
+      .map(rule -> rule.validate(quickMarc.getLeader()))
+      .filter(Optional::isPresent)
+      .map(Optional::get)
+      .collect(Collectors.toList()));
+     
     if (validationErrors.isEmpty()) {
       return new ValidationResult(true, Collections.emptyList());
     } else {